дополнения к беку
