+++ utils/constants.js
@@ -1,0 +1,10 @@
+const allowedCors = [
+  'https://movies-explorer-pgk.nomoredomains.xyz',
+  'http://movies-explorer-pgk.nomoredomains.xyz',
+  'http://localhost:3001',
+  'http://localhost:3000',
+];
+
+module.exports = {
+  allowedCors,
+};
